Take advantage of fMBT 0.2.2 features
Use adapter_init instead of preprocessor to run certain code only when
the AAL module is used as an adapter. This code is not executed when
fmbt-editor simulates test execution.

Use shorthand adapter="aal" to use the same AAL module as an adapter
that has been configured to be used as a model, too.